+++ src/main/java/com/yeahn/plan/dao/PlanMapper.java
@@ -11,10 +11,10 @@ public interface PlanMapper {
     List<PlanVo> getPlanList(PlanVo planVo);
 
     // 운동 계획 상세 조회 (Master)
-    PlanVo getPlan(Integer planSeq);
+    PlanVo getPlan(Long planSeq);
 
     // 운동 계획 상세 목록 조회 (Detail)
-    List<PlanDetailVo> getPlanDetails(Integer planSeq);
+    List<PlanDetailVo> getPlanDetails(Long planSeq);
 
     // 운동 계획 마스터 저장
     int insertPlan(PlanVo planVo);
@@ -33,5 +33,5 @@ public interface PlanMapper {
 
     // 특정 계획의 상세 목록 삭제 (물리 삭제 또는 Soft Delete)
     // 일반적으로 대량 수정 시 상세를 전체 삭제 후 재등록하는 방식이 편리합니다.
-    int deletePlanDetailsByPlanSeq(Integer planSeq);
+    int deletePlanDetailsByPlanSeq(Long planSeq);
 }
